CAS: Split up BuiltinCAS.cpp, NFC
Split the in-memory and on-disk builtin CAS implementations out of
BuiltinCAS.cpp into InMemoryCAS.cpp and OnDiskCAS.cpp, respectively,
and split out a BuiltinCAS.h header.
