+++ llvm/lib/CAS/CMakeLists.txt
@@ -7,8 +7,10 @@ add_llvm_component_library(LLVMCAS
   FileSystemCache.cpp
   HashMappedTrie.cpp
   HierarchicalTreeBuilder.cpp
+  InMemoryCAS.cpp
   LazyMappedFileRegion.cpp
   LazyMappedFileRegionBumpPtr.cpp
+  OnDiskCAS.cpp
   OnDiskHashMappedTrie.cpp
   PluginCAS.cpp
   Utils.cpp
